#1954: cleanup: * fix typos * extract test helper TestStreamConsumer to avoid polluting InMemoryConsumer * remove unused AgentInstance#supportsBuildCommandProtocol
